requirements.txt
Django==2.1.3
mysqlclient==1.3.13
wheel==0.32.3
gunicorn==19.9.0
psycopg2==2.7.6.1
djangorestframework==3.9.0

